Public Documentation

Documentation for MimiqCircuits.jl's public interface.

See the Internals section of the manual for internal package docs.

Modules

MimiqCircuits.execute — Method

execute(connection, circuit[; kwargs...])

Execute a quantum circuit on the MIMIQ remote services.

The circuit is applied to the zero state and the resulting state is measured via sampling. Optionally amplitudes corresponding to few selected bit states (or bitstrings) can be returned from the computation.

Keyword Arguments

label::String: mnemonic name to give to the simulation, will be visible on the web interface

algorithm: algorithm to use by the compuation. By default "auto" will select the fastest algorithm between "statevector" or "mps".

nsamples::Integer: number of times to sample the circuit (default: 1000, maximum: 2^16)

bitstates::Vector{BitState}: list of bit states to compute the amplitudes for (default: BitState[])

timelimit: number of seconds before the computation is stopped (default: 300 seconds or 5 minutes)

bonddim::Int64: bond dimension for the MPS algorithm (default: 256, maximum: 4096)

seed::Int64: a seed for running the simulation (default: random seed)

MimiqCircuits.getinputs — Method

getinputs(connection, execution)

Returns the circuit and parameters for the given execution.

MimiqCircuits.getresults — Method

getresults(connection, execution; kwargs...)

Block until the given execution is finished and return the results.

Keyword Arguments

interval: time interval in seconds to check for job completion (default: 10)

MimiqCircuits.printreport — Method

printreport(res::MimiqCircuits.Results; kwargs...)

Print a report on the MIMIQ simulation results res

Keyword Arguments

max_outcomes: the maximum number of unique measurement outcomes to display (default 8)

MimiqLink.MimiqLink — Module

module MimiqLink end

This module contains convenience tools to establish and keep up a connection to the QPerfect MIMIQ services, both remote or on premises.

It allows for three different connection modes: via login page, via token, via credentials

Login Page

This method will open a browser pointing to a login page. The user will be asked to insert username/email and password.

julia> using MimiqLink

julia> connection = MimiqLink.connect()

optionally an address for the MIMIQ services can be specified

julia> connection = MimiqLink.connect(uri = "http://127.0.0.1/api")

Token

This method will allow the user to save a token file (by login via a login page), and then load it also from another julia session.

julia> using MimiqLink

julia> MimiqLink.savetoken(uri = "http://127.0.0.1/api")

this will save a token in the qperfect.json file in the current directory. In another julia session is then possible to do:

julia> using MimiqLink

julia> connection = MimiqLink.loadtoken("path/to/my/qperfect.json")

Credentials

This method will allow users to access by directly use their own credentials.

WARNING it is strongly discuraged to use this method. If files with credentials will be shared the access to the qperfect account might be compromised.

julia> using MimiqLink

julia> connection = MimiqLink.connect("me@mymail.com", "myweakpassword")

julia> MimiqLink.connect("me@mymail.com", "myweakpassword"; uri = "http://127.0.0.1/api")

Quantum Circuits and Instructions

MimiqCircuitsBase.Instruction — Type

struct Instruction{N,M,T<:Operation}

Element of a quantum circuit, representing a N-qubit gate applied to N targets

Parameters

gate::T actual gate represented

qtargets::NTuple{N, Int64} indices specifying the quantum bits on which the instruction is applied

ctargets::NTuple{N, Int64} indices specifying the classical bits on which the instruction is applied

MimiqCircuitsBase.getbit — Function

getbit(instruction, i)

Returns the i-th target classical bit of an instruction.

See also getbits, getqubit, getqubits,

MimiqCircuitsBase.getbits — Function

getbits(instruction)

Returns all the classical bits to which the instruction is applied.

See also getbit, getqubits, getqubit,

MimiqCircuitsBase.getoperation — Method

getoperation(getoperation)

Returns the quantum operation associated to the given gate instruction.

MimiqCircuitsBase.getqubit — Function

getqubit(instruction, i)

Returns the i-th target qubit of an instruction.

See also getqubits, getbit, getbits,

MimiqCircuitsBase.getqubits — Function

getqubits(instruction)

Returns all the quantum bits to which the instruction is applied.

See also getqubit, getbits, getbit,

MimiqCircuitsBase.gettarget — Function

gettarget(instruction, i)

Returns the i-th target qubit of an instruction.

!!!warn Deprecated in favor of getqubit and getbit

MimiqCircuitsBase.gettargets — Function

gettargets(instruction)

Returns all the quantum qubits to which the instruction is applied.

!!!warn Deprecated in favor of getqubits and getbits

MimiqCircuitsBase.Circuit — Type

struct Circuit

Representation of a quantum circuit as a vector of gates applied to the qubits.

Parameters

gates::Vector{Instruction} vector of quantum instructions (see Instruction)

Example iteration

circuit = Circuit()
# add gates to circuit

for (; operation, targets) in circuit
    # do something with the gate and its targets
    # e.g.
end

(here the iteration parameters should be called operation and targets to proper destructure a Instruction)

Gate types

Single qubit gates (basic): GateX, GateY, GateZ, GateH, GateS, GateSDG, GateT, GateTDG, GateSX, GateSXDG, GateID

Single qubit gates (parametric): GateRX, GateRY, GateRZ, GateP, GateR, GateU

Two qubit gates (basic): GateCX, GateCY, GateCZ, GateCH, GateSWAP, GateISWAP, GateISWAPDG

Two qubit gates (parametric): GateCP, GateCRX, GateCRY, GateCRZ, GateCU

Custom gate (currently only for 1 and 2 qubit gates): GateCustom

Special operations

Barrier, Reset, Measure

Operations

Gates

Single qubit gates

MimiqCircuitsBase.GateH — Type

struct GateH <: Gate{1}

Single qubit Hadamard gate.

Matrix Representation

\[\operatorname H = \frac{1}{\sqrt{2}} \begin{pmatrix} 1 & 1 \\ 1 & -1 \end{pmatrix}\]

Examples

julia> matrix(GateH())
2×2 Matrix{Float64}:
 0.707107   0.707107
 0.707107  -0.707107

julia> push!(Circuit(), GateH(), 1)
1-qubit circuit with 1 gates:
└── H @ q1

MimiqCircuitsBase.GateID — Type

struct GateID <: Gate{1}

Single qubit Identity gate

Matrix Representation

\[\operatorname{I} = \begin{pmatrix} 1 & 0 \\ 0 & 1 \end{pmatrix}\]

Examples

julia> matrix(GateID())
2×2 Matrix{Float64}:
 1.0  -0.0
 0.0   1.0

julia> push!(Circuit(), GateID(), 1)
1-qubit circuit with 1 gates:
└── ID @ q1

MimiqCircuitsBase.GateS — Type

struct GateS <: Gate{1}

Single qubit S gate (or Phase gate).

See also GateSDG

Matrix Representation

\[\operatorname S = \begin{pmatrix} 1 & 0 \\ 0 & i \end{pmatrix}\]

Examples

julia> matrix(GateS())
2×2 Matrix{ComplexF64}:
 1.0+0.0im  0.0+0.0im
 0.0+0.0im  0.0+1.0im

julia> push!(Circuit(), GateS(), 1)
1-qubit circuit with 1 gates:
└── S @ q1

MimiqCircuitsBase.GateSDG — Type

struct GateSDG <: Gate{1}

Single qubit S-dagger gate (conjugate transpose of the S gate).

See also GateS

Matrix Representation

\[\operatorname S^\dagger = \begin{pmatrix} 1 & 0 \\ 0 & -i \end{pmatrix}\]

Examples

julia> matrix(GateSDG())
2×2 Matrix{ComplexF64}:
 1.0+0.0im  0.0+0.0im
 0.0+0.0im  0.0-1.0im

julia> push!(Circuit(), GateSDG(), 1)
1-qubit circuit with 1 gates:
└── SDG @ q1

MimiqCircuitsBase.GateSX — Type

struct GateSX <: Gate{1}

Single qubit √X gate.

See also GateSXDG, GateX

Matrix Representation

\[\sqrt{\operatorname{X}} = \frac{1}{2} \begin{pmatrix} 1+i & 1-i \\ 1-i & 1+i \end{pmatrix}\]

Examples

julia> matrix(GateSX())
2×2 Matrix{ComplexF64}:
 0.5+0.5im  0.5-0.5im
 0.5-0.5im  0.5+0.5im

julia> push!(Circuit(), GateSX(), 1)
1-qubit circuit with 1 gates:
└── SX @ q1

MimiqCircuitsBase.GateSXDG — Type

struct GateSXDG <: Gate{1}

Single qubit √X-dagger gate (conjugate transpose of the √X gate)

See also GateSX, GateX

Matrix Representation

\[\sqrt{\operatorname{X}}^\dagger = \frac{1}{2} \begin{pmatrix} 1-i & 1+i \\ 1+i & 1-i \end{pmatrix}\]

Examples

julia> matrix(GateSXDG())
2×2 Matrix{ComplexF64}:
 0.5-0.5im  0.5+0.5im
 0.5+0.5im  0.5-0.5im

julia> push!(Circuit(), GateSXDG(), 1)
1-qubit circuit with 1 gates:
└── SXDG @ q1

MimiqCircuitsBase.GateT — Type

struct GateT <: Gate{1}

Single qubit T gate.

See also GateTDG

Matrix Representation

\[\operatorname T = \begin{pmatrix} 1 & 0 \\ 0 & \exp(\frac{i\pi}{4}) \end{pmatrix}\]

Examples

julia> matrix(GateT())
2×2 Matrix{ComplexF64}:
 1.0+0.0im       0.0+0.0im
 0.0+0.0im  0.707107+0.707107im

julia> push!(Circuit(), GateT(), 1)
1-qubit circuit with 1 gates:
└── T @ q1

MimiqCircuitsBase.GateTDG — Type

struct GateTDG <: Gate{1}

Single qubit T-dagger gate (conjugate transpose of the T gate).

See also GateT

Matrix Representation

\[\operatorname T^\dagger = \begin{pmatrix} 1 & 0 \\ 0 & \exp(\frac{-i\pi}{4}) \end{pmatrix}\]

Examples

julia> matrix(GateTDG())
2×2 Matrix{ComplexF64}:
 1.0+0.0im       0.0+0.0im
 0.0+0.0im  0.707107-0.707107im

julia> push!(Circuit(), GateTDG(), 1)
1-qubit circuit with 1 gates:
└── TDG @ q1

MimiqCircuitsBase.GateX — Type

struct GateX <: Gate{1}

Single qubit Pauli-X gate.

Matrix Representation

\[\operatorname X = \begin{pmatrix} 0 & 1 \\ 1 & 0 \end{pmatrix}\]

Examples

julia> matrix(GateX())
2×2 Matrix{Float64}:
 0.0   1.0
 1.0  -0.0

julia> push!(Circuit(), GateX(), 1)
1-qubit circuit with 1 gates:
└── X @ q1

MimiqCircuitsBase.GateY — Type

struct GateY <: Gate{1}

Single qubit Pauli-Y gate.

Matrix Representation

\[\operatorname Y = \begin{pmatrix} 0 & -i \\ i & 0 \end{pmatrix}\]

Examples

julia> matrix(GateY())
2×2 Matrix{ComplexF64}:
 0.0+0.0im  -0.0-1.0im
 0.0+1.0im  -0.0+0.0im

julia> push!(Circuit(), GateY(), 1)
1-qubit circuit with 1 gates:
└── Y @ q1

MimiqCircuitsBase.GateZ — Type

struct GateZ <: Gate{1}

Single qubit Pauli-Z gate.

Matrix Representation

\[\operatorname Z = \begin{pmatrix} 1 & 0 \\ 0 & -1 \end{pmatrix}\]

Examples

julia> matrix(GateZ())
2×2 Matrix{Float64}:
 1.0   0.0
 0.0  -1.0

julia> push!(Circuit(), GateZ(), 1)
1-qubit circuit with 1 gates:
└── Z @ q1

Two-qubit gates

MimiqCircuitsBase.GateCH — Type

struct GateCH <: Gate{2}

Two qubit Controlled-Hadamard gate.

Matrix Representation

\[\operatorname{CH} = \begin{pmatrix} 1 & 0 & 0 & 0 \\ 0 & 1 & 0 & 0 \\ 0 & 0 & \frac{1}{\sqrt{2}} & \frac{1}{\sqrt{2}} \\ 0 & 0 & \frac{1}{\sqrt{2}} & -\frac{1}{\sqrt{2}} \end{pmatrix}\]

By convention we refer to the first qubit as the control qubit and the second qubit as the target.

Examples

julia> matrix(GateCH())
4×4 Matrix{Float64}:
 1.0  0.0  0.0        0.0
 0.0  1.0  0.0        0.0
 0.0  0.0  0.707107   0.707107
 0.0  0.0  0.707107  -0.707107

julia> push!(Circuit(), GateCH(), 1, 2)
2-qubit circuit with 1 gates:
└── CH @ q1, q2

MimiqCircuitsBase.GateCS — Type

struct GateCS <: Gate{2}

Two qubit Controlled-S gate.

Matrix Representation

\[\begin{pmatrix} 1 & 0 & 0 & 0 \\ 0 & 1 & 0 & 0 \\ 0 & 0 & 1 & 0 \\ 0 & 0 & 0 & i \end{pmatrix}\]

Examples

julia> matrix(GateCS())
4×4 Matrix{ComplexF64}:
 1.0+0.0im  0.0+0.0im  0.0+0.0im  0.0+0.0im
 0.0+0.0im  1.0+0.0im  0.0+0.0im  0.0+0.0im
 0.0+0.0im  0.0+0.0im  1.0+0.0im  0.0+0.0im
 0.0+0.0im  0.0+0.0im  0.0+0.0im  0.0+1.0im

julia> push!(Circuit(), GateCS(), 1, 2)
2-qubit circuit with 1 gates:
└── CS @ q1, q2

MimiqCircuitsBase.GateCSDG — Type

struct GateCSDG <: Gate{2}

Two qubit CS-dagger gate.

Matrix Representation

\[\begin{pmatrix} 1 & 0 & 0 & 0 \\ 0 & 1 & 0 & 0 \\ 0 & 0 & 1 & 0 \\ 0 & 0 & 0 & i \end{pmatrix}\]

Examples

julia> matrix(GateCSDG())
4×4 Matrix{ComplexF64}:
 1.0+0.0im  0.0+0.0im  0.0+0.0im  0.0+0.0im
 0.0+0.0im  1.0+0.0im  0.0+0.0im  0.0+0.0im
 0.0+0.0im  0.0+0.0im  1.0+0.0im  0.0+0.0im
 0.0+0.0im  0.0+0.0im  0.0+0.0im  0.0-1.0im

julia> push!(Circuit(), GateCSDG(), 1, 2)
2-qubit circuit with 1 gates:
└── CSDG @ q1, q2

MimiqCircuitsBase.GateCSX — Type

struct GateCSX <: Gate{2}

Two qubit Controlled-SX gate. (Control on second qubit)

Matrix Representation

\[\begin{pmatrix} 1 & 0 & 0 & 0 \\ 0 & \frac{1+i}{\sqrt{2}} & 0 & \frac{1-i}{\sqrt{2}} \\ 0 & 0 & 1 & 0 \\ 0 & \frac{1-i}{\sqrt{2}} & 0 & \frac{1+i}{\sqrt{2}} \end{pmatrix}\]

Examples

julia> matrix(GateCSX())
4×4 Matrix{ComplexF64}:
 1.0+0.0im  0.0+0.0im  0.0+0.0im  0.0+0.0im
 0.0+0.0im  0.5+0.5im  0.0+0.0im  0.5-0.5im
 0.0+0.0im  0.0+0.0im  1.0+0.0im  0.0+0.0im
 0.0+0.0im  0.5-0.5im  0.0+0.0im  0.5+0.5im

julia> push!(Circuit(), GateCSX(), 1, 2)
2-qubit circuit with 1 gates:
└── CSX @ q1, q2

MimiqCircuitsBase.GateCSXDG — Type

struct GateCSXDG <: Gate{2}

Two qubit CSX-dagger gate. (Control on second qubit)

Matrix Representation

\[\begin{pmatrix} 1 & 0 & 0 & 0 \\ 0 & \frac{1-i}{\sqrt{2}} & 0 & \frac{1+i}{\sqrt{2}} \\ 0 & 0 & 1 & 0 \\ 0 & \frac{1+i}{\sqrt{2}} & 0 & \frac{1-i}{\sqrt{2}} \end{pmatrix}\]

Examples

julia> matrix(GateCSXDG())
4×4 Matrix{ComplexF64}:
 1.0+0.0im  0.0+0.0im  0.0+0.0im  0.0+0.0im
 0.0+0.0im  0.5-0.5im  0.0+0.0im  0.5+0.5im
 0.0+0.0im  0.0+0.0im  1.0+0.0im  0.0+0.0im
 0.0+0.0im  0.5+0.5im  0.0+0.0im  0.5-0.5im

julia> push!(Circuit(), GateCSXDG(), 1, 2)
2-qubit circuit with 1 gates:
└── CSXDG @ q1, q2

MimiqCircuitsBase.GateCX — Type

struct GateCX <: Gate{2}

Two qubit Controlled-X gate (or CNOT).

Matrix Representation

\[\operatorname{CX} = \begin{pmatrix} 1 & 0 & 0 & 0 \\ 0 & 1 & 0 & 0 \\ 0 & 0 & 0 & 1 \\ 0 & 0 & 1 & 0 \end{pmatrix}\]

By convention we refer to the first qubit as the control qubit and the second qubit as the target.

Examples

julia> matrix(GateCX())
4×4 Matrix{Float64}:
 1.0  0.0  0.0   0.0
 0.0  1.0  0.0   0.0
 0.0  0.0  0.0   1.0
 0.0  0.0  1.0  -0.0

julia> push!(Circuit(), GateCX(), 1, 2)
2-qubit circuit with 1 gates:
└── CX @ q1, q2

MimiqCircuitsBase.GateCY — Type

struct GateCY <: Gate{2}

Two qubit Controlled-Y gate.

Matrix Representation

\[\operatorname{CY} = \begin{pmatrix} 1 & 0 & 0 & 0 \\ 0 & 1 & 0 & 0 \\ 0 & 0 & 0 & -i \\ 0 & 0 & i & 0 \end{pmatrix}\]

By convention we refer to the first qubit as the control qubit and the second qubit as the target.

Examples

julia> matrix(GateCY())
4×4 Matrix{ComplexF64}:
 1.0+0.0im  0.0+0.0im  0.0+0.0im   0.0+0.0im
 0.0+0.0im  1.0+0.0im  0.0+0.0im   0.0+0.0im
 0.0+0.0im  0.0+0.0im  0.0+0.0im  -0.0-1.0im
 0.0+0.0im  0.0+0.0im  0.0+1.0im  -0.0+0.0im

julia> push!(Circuit(), GateCY(), 1, 2)
2-qubit circuit with 1 gates:
└── CY @ q1, q2

MimiqCircuitsBase.GateCZ — Type

struct GateCZ <: Gate{2}

Two qubit Controlled-Z gate.

Matrix Representation

\[\operatorname{CZ} = \begin{pmatrix} 1 & 0 & 0 & 0 \\ 0 & 1 & 0 & 0 \\ 0 & 0 & 1 & 0 \\ 0 & 0 & 0 & -1 \end{pmatrix}\]

By convention we refer to the first qubit as the control qubit and the second qubit as the target.

Examples

julia> matrix(GateCZ())
4×4 Matrix{Float64}:
 1.0  0.0  0.0   0.0
 0.0  1.0  0.0   0.0
 0.0  0.0  1.0   0.0
 0.0  0.0  0.0  -1.0

julia> push!(Circuit(), GateCZ(), 1, 2)
2-qubit circuit with 1 gates:
└── CZ @ q1, q2

MimiqCircuitsBase.GateDCX — Type

struct GateDCX <: Gate{2}

Two qubit double-CNOT (Control on first qubit and then second) OR DCX gate.

Matrix Representation

\[\begin{pmatrix} 1 & 0 & 0 & 0 \\ 0 & 0 & 0 & 1 \\ 0 & 1 & 0 & 0 \\ 0 & 0 & 1 & 0 \end{pmatrix}\]

Examples

julia> matrix(GateDCX())
4×4 Matrix{Float64}:
 1.0  0.0  0.0  0.0
 0.0  0.0  0.0  1.0
 0.0  1.0  0.0  0.0
 0.0  0.0  1.0  0.0

julia> push!(Circuit(), GateDCX(), 1, 2)
2-qubit circuit with 1 gates:
└── DCX @ q1, q2

MimiqCircuitsBase.GateDCXDG — Type

struct GateDCXDG <: Gate{2}

Two qubit DCX-dagger gate.

Matrix Representation

\[\begin{pmatrix} 1 & 0 & 0 & 0 \\ 0 & 0 & 1 & 0 \\ 0 & 0 & 0 & 1 \\ 0 & 1 & 0 & 0 \end{pmatrix}\]

Examples

julia> matrix(GateDCXDG())
4×4 Matrix{Float64}:
 1.0  0.0  0.0  0.0
 0.0  0.0  1.0  0.0
 0.0  0.0  0.0  1.0
 0.0  1.0  0.0  0.0

julia> push!(Circuit(), GateDCXDG(), 1, 2)
2-qubit circuit with 1 gates:
└── DCXDG @ q1, q2

MimiqCircuitsBase.GateECR — Type

struct GateECR <: Gate{2}

Two qubit ECR echo gate.

Matrix Representation

\[\begin{pmatrix} 0 & \frac{1}{\sqrt{2}} \ & 0 & \frac{i}{\sqrt{2}} \\ \frac{1}{\sqrt{2}} & 0 & \frac{-i}{\\sqrt{2}} & 0 \\ 0 & \frac{i}{\\sqrt{2}} & 0 & \frac{i}{\sqrt{2}} \\ \frac{-i}{\sqrt{2}} & 0 & \frac{1}{\sqrt{2}} & 0 \end{pmatrix}\]

Examples

julia> matrix(GateCSX())
4×4 Matrix{ComplexF64}:
 1.0+0.0im  0.0+0.0im  0.0+0.0im  0.0+0.0im
 0.0+0.0im  0.5+0.5im  0.0+0.0im  0.5-0.5im
 0.0+0.0im  0.0+0.0im  1.0+0.0im  0.0+0.0im
 0.0+0.0im  0.5-0.5im  0.0+0.0im  0.5+0.5im

julia> push!(Circuit(), GateCSX(), 1, 2)
2-qubit circuit with 1 gates:
└── CSX @ q1, q2

MimiqCircuitsBase.GateISWAP — Type

struct GateISWAP <: Gate{2}

Two qubit ISWAP gate.

See also GateISWAPDG, GateSWAP.

Matrix Representation

\[\operatorname{ISWAP} = \frac{1}{\sqrt{2}} \begin{pmatrix} 1 & 0 & 0 & 0 \\ 0 & 0 & i & 0 \\ 0 & i & 0 & 0 \\ 0 & 0 & 0 & 1 \end{pmatrix}\]

Examples

julia> matrix(GateISWAP())
4×4 Matrix{ComplexF64}:
 1.0+0.0im  0.0+0.0im  0.0+0.0im  0.0+0.0im
 0.0+0.0im  0.0+0.0im  0.0+1.0im  0.0+0.0im
 0.0+0.0im  0.0+1.0im  0.0+0.0im  0.0+0.0im
 0.0+0.0im  0.0+0.0im  0.0+0.0im  1.0+0.0im

julia> push!(Circuit(), GateISWAP(), 1, 2)
2-qubit circuit with 1 gates:
└── ISWAP @ q1, q2

MimiqCircuitsBase.GateISWAPDG — Type

struct GateISWAPDG <: Gate{2}

Two qubit ISWAP-dagger gate (conjugate transpose of ISWAP)

See also GateISWAP, GateSWAP

Matrix Representation

\[\operatorname{ISWAP}^\dagger = \frac{1}{\sqrt{2}} \begin{pmatrix} 1 & 0 & 0 & 0 \\ 0 & 0 & -i & 0 \\ 0 & -i & 0 & 0 \\ 0 & 0 & 0 & 1 \end{pmatrix}\]

Examples

julia> matrix(GateISWAPDG())
4×4 Matrix{ComplexF64}:
 1.0+0.0im  0.0+0.0im  0.0+0.0im  0.0+0.0im
 0.0+0.0im  0.0+0.0im  0.0-1.0im  0.0+0.0im
 0.0+0.0im  0.0-1.0im  0.0+0.0im  0.0+0.0im
 0.0+0.0im  0.0+0.0im  0.0+0.0im  1.0+0.0im

julia> push!(Circuit(), GateISWAPDG(), 1, 2)
2-qubit circuit with 1 gates:
└── ISWAPDG @ q1, q2

MimiqCircuitsBase.GateSWAP — Type

struct GateSWAP <: Gate{2}

Two qubit SWAP gate.

See also GateISWAP

Matrix Representation

\[\operatorname{SWAP} = \frac{1}{\sqrt{2}} \begin{pmatrix} 1 & 0 & 0 & 0 \\ 0 & 0 & 1 & 0 \\ 0 & 1 & 0 & 0 \\ 0 & 0 & 0 & 1 \end{pmatrix}\]

Examples

julia> matrix(GateSWAP())
4×4 Matrix{Float64}:
 1.0  0.0  0.0  0.0
 0.0  0.0  1.0  0.0
 0.0  1.0  0.0  0.0
 0.0  0.0  0.0  1.0

julia> push!(Circuit(), GateSWAP(), 1, 2)
2-qubit circuit with 1 gates:
└── SWAP @ q1, q2

Multi-qubit

MimiqCircuitsBase.GateCCX — Type

struct GateCCX <: Gate{3}

C₂X (or C₂NOT) 3-qubits gate. Where the first two qubits are used as controls.

MimiqCircuitsBase.GateCSWAP — Type

struct GateCSWAP <: Gate{3} end

3-qubits control SWAP gate where the first qubit is the control.

Custom gates

MimiqCircuitsBase.GateCustom — Type

struct GateCustom{N,T} <: Gate{N}

N qubit gate specified by a $2^N \times 2^N$ matrix with elements of type T.

Use this to construct your own gates based on unitary matrices. Currently only N=1,2 (M=2,4) are recognised.

MIMIQ uses textbook convention for specifying gates.

One qubit gate matrices are defined in the basis $|0\rangle$, $|1\rangle$ e.g.,

\[\operatorname{Z} = \begin{pmatrix} 1&0\\ 0&-1 \end{pmatrix}\]

Two qubit gate matrices are defined in the basis $|00\rangle$, $|01\rangle$>, $|10\rangle$, $|11\rangle$ where the left-most qubit is the first to appear in the target list e.g.,

\[\operatorname{CNOT} = \begin{pmatrix} 1&0&0&0\\ 0&1&0&0\\ 0&0&0&1\\ 0&0&1&0 \end{pmatrix}\]

julia> CNOT = [1 0 0 0; 0 1 0 0; 0 0 0 1; 0 0 1 0]
4×4 Matrix{Int64}:
 1  0  0  0
 0  1  0  0
 0  0  0  1
 0  0  1  0

julia> # CNOT gate with control on q1 and target on q2

julia> Instruction(GateCustom(CNOT), 1, 2)
GateCustom([1 0 0 0; 0 1 0 0; 0 0 0 1; 0 0 1 0]) @ q1, q2

# Examples

jldoctest julia> g = GateCustom([1 0; 0 1]) Custom([1.0 0.0; 0.0 1.0])

julia> push!(Circuit(), g, 1) 1-qubit circuit with 1 gates: └── Custom([1.0 0.0; 0.0 1.0]) @ q1 ```

Non-unitary operations

MimiqCircuitsBase.Reset — Type

struct Reset <: Operation{1}

Quantum operation that resets the status of one qubit to the $\ket{0}$ state.

Examples

julia> push!(Circuit(), Reset(), 1)
2-qubit circuit with 1 gates:
└── Reset @ q1

MimiqCircuitsBase.Measure — Type

struct Measure <: Operation{1}

Single qubit measurement operation in the computational basis

This operation is non-reversible

Examples

Measure project the qubit state and optionally store the result of the measurement for that qubit in a classical register.

To just apply the measurement on qubit 1 and discard the result, do:

julia> push!(Circuit(), Measure(), 1)
2-qubit circuit with 1 gates:
└── Measure @ q1

In order to store the result on the 2nd bit, call:

julia> push!(Circuit(), Measure(), 1 => 2)
2-qubit circuit with 1 gates:
└── Measure @ q1, c1

No-ops

MimiqCircuitsBase.Barrier — Type

struct Barrier <: Operation{1}

A barrier is a special operation that does not affect the quantum state or the execution of a circuit, but it prevents compression or optimization operation from being applied across it.

Examples

julia> push!(Circuit(), Barrier(), 1)
2-qubit circuit with 1 gates:
└── Barrier @ q1

julia> push!(Circuit(), Barrier(), 1:4...)
2-qubit circuit with 1 gates:
└── Barrier @ q1, q2, q3, q4

julia> push!(Circuit(), Barrier(), 1:4)
2-qubit circuit with 1 gates:
├── Barrier @ q1
├── Barrier @ q2
├── Barrier @ q3
└── Barrier @ q4

Composite operations

MimiqCircuitsBase.IfStatement — Type

struct IfStatement{N,M} <: Operation{N,M}

MimiqCircuitsBase.Control — Type

struct Control{N,M,L,T<:Operation{M,0}} <: Operation{L,0} end

MimiqCircuitsBase.Parallel — Type

struct Parallel{N,M,L,T<:Operation{M,0}} <: Operation{L,0} end

Bit States

MimiqCircuitsBase.BitState — Type

struct BitState

Representation of the quantum state of a quantum register with definite values for each qubit.

Examples

julia> BitState(16)
16-qubit BitState with 0 non-zero qubits:
└── |0000000000000000⟩

julia> bs = BitState(16, [1,2,3,4])
16-qubit BitState with 4 non-zero qubits:
├── |1111000000000000⟩
└── non-zero qubits: [1, 2, 3, 4]

julia> bs[10] = 1
1

julia> bs
16-qubit BitState with 5 non-zero qubits:
├── |1111000001000000⟩
└── non-zero qubits: [1, 2, 3, 4, 10]

julia> c = Circuit()
empty circuit

julia> push!(c, GateX(), 8)
8-qubit circuit with 1 gates:
└── X @ q8

julia> BitState(c, [1,3,5,8])
8-qubit BitState with 4 non-zero qubits:
├── |10101001⟩
└── non-zero qubits: [1, 3, 5, 8]

julia> bitstate_to_integer(bs)
527

julia> typeof(ans)
BigInt

julia> bitstate_to_integer(bs, Int64)
527

julia> typeof(ans)
Int64

There are many different ways to get bit states:

julia> bs = BitState(30, 2344574)
30-qubit BitState with 13 non-zero qubits:
├── |011111100110001111000100000000⟩
└── non-zero qubits: [2, 3, 4, 5, 6, 7, 10, 11, 15, 16, 17, 18, 22]

julia> ones(BitState, 10) # or also trues(BitState, 10)
10-qubit BitState with 10 non-zero qubits:
├── |1111111111⟩
└── non-zero qubits: [1, 2, 3, 4, 5, 6, 7, 8, 9, 10]

julia> zeros(BitState, 10) # or also falses(BitState, 10)
10-qubit BitState with 0 non-zero qubits:
└── |0000000000⟩

julia> BitState(16) do i
           iseven(i)
       end
16-qubit BitState with 8 non-zero qubits:
├── |0101010101010101⟩
└── non-zero qubits: [2, 4, 6, 8, 10, 12, 14, 16]

MimiqCircuitsBase.bits — Method

bits(bitstate)

Get the underlying bit array of a bit state.

MimiqCircuitsBase.bitstate_to_index — Method

bitstate_to_index(bitstate)

Convert a bit state into the corresponding index.

This is useful for indexing, for example, a vector of states.

MimiqCircuitsBase.bitstate_to_integer — Method

bitstate_to_integer(bitstate[, T])

Convert a bit state into its corresponding integer.

MimiqCircuitsBase.nonzeros — Method

nonzeros(bitstate)

Return the indices of the non-zero qubits in a bit state.

MimiqCircuitsBase.to01 — Method

to01(bitstate[, endianess=:big])

Converts a BitState into a string of 0 and 1 characters. Optionally endianess can be specified, which can be either :big or :little.

Examples

julia> to01(bs"10011")
"10011"

julia> to01(bs"10011"; endianess=:big)
"10011"

julia> to01(bs"10011"; endianess=:little)
"11001"

MimiqCircuitsBase.@bs_str — Macro

macro bs_str(s)

Convert a string into a bit state.

Examples

julia> bs"101011"
6-qubit BitState with 4 non-zero qubits:
├── |101011⟩
└── non-zero qubits: [1, 3, 5, 6]