Private types and functions

MimiqCircuitsBase.LazyArg — Type

LazyArg()

Placeholder for a lazy argument in a LazyExpr.

MimiqCircuitsBase.LazyExpr — Type

LazyExpr(type, args)

Helps evaluating expressions lazily.

Evaluation occurs only then the LazyExpr is called with some arguments, and the arguments will be passed to the inner part of the expression.

MimiqCircuitsBase.UndefinedParameterError — Type

UndefinedParameterError(name, gatename)

The parameters of the gate are undefined. name is the name of the parameter, gatename the one of the gate.

MimiqCircuitsBase.evaluate! — Function

evaluate!(circutit, dictionary)

Evaluate the

MimiqCircuitsBase.exponent — Method

exponent(poweroperation)

Exponent associated with a power operation

Examples

julia> MimiqCircuitsBase.exponent(power(GateH(), 2))
2

julia> MimiqCircuitsBase.exponent(GateSX())
1//2

MimiqCircuitsBase.unwrappedmatrix — Method

unwrappedmatrix(gate)

Returns the matrix associated to the specified quantum gate without the Symbolics.Num wrapper.

Note

If any of the gate's parameters is symbolic, an error is thrown.

See also matrix.

Examples

julia> unwrappedmatrix(GateRX(π/2))
2×2 Matrix{ComplexF64}:
    0.707107+5.55112e-17im       0.0-0.707107im
 1.11022e-16-0.707107im     0.707107+5.55112e-17im

julia> unwrappedmatrix(GateH())
2×2 Matrix{Float64}:
 0.707107   0.707107
 0.707107  -0.707107

MimiqLink.DEFAULT_INTERVAL — Constant

const DEFAULT_INTERVAL

Default refresh interval for tokens (in seconds)

MimiqLink.Connection — Type

struct Connection

Connection with the MIMIQ Services.

Attributes

uri: the URI of the connected instance

tokens_channel: channel updated with the latest refreshed token

refresher: task that refreshes the token on a configured interval

MimiqLink.Execution — Type

struct Execution

Structure referring to an execution on the MIMIQ Services.

MimiqLink.Tokens — Type

struct Tokens

Store access and refresh tokens

MimiqLink.refresh — Method

refresh(tokens, uri)

Refresh the tokens at the given uri / instance of MIMIQ.