Private types and functions

Documentation for MimiqCircuits.jl's internal interface.

MimiqCircuitsBase.fromdict — Function

fromdict(object, dict)

Returns the given object from the serialized JSON dictionary.

MimiqCircuitsBase.todict — Function

todict(circuit)
todict(instruction)
todict(operation)

MimiqLink.DEFAULT_INTERVAL — Constant

const DEFAULT_INTERVAL

Default refresh interval for tokens (in seconds)

MimiqLink.QPERFECT_CLOUD — Constant

const QPERFECT_CLOUD

Fallback address for the QPerfect Cloud services

MimiqLink.Connection — Type

struct Connection

Connection with the MIMIQ Services.

Attributes

uri: the URI of the connected instance

tokens_channel: channel updated with the latest refreshed token

refresher: task that refreshes the token on a configured interval

MimiqLink.Execution — Type

struct Execution

Structure referring to an execution on the MIMIQ Services.

MimiqLink.Tokens — Type

struct Tokens

Store access and refresh tokens

MimiqLink.connect — Function

connect([; url=https://mimiq.qperfect.io])
connect(token[; url=https://mimiq.qperfect.io])
connect(username, password[; url=https://mimiq.qperfect.io])

Establish a connection to the MIMIQ Services.

A refresh process will be spawned in the background to refresh the access credentials. An active connection can be closed by using the close(connection) method. As an example:

connection = connect("john.doe@example.com", "johnspassword")
close(connection)

Warning

The first method will open a login page in the default browser and ask for your email and password. This method is encouraged, as it will avoid saving your password as plain text in your scripts or notebooks.

MimiqLink.loadtoken — Function

loadtoken([filename])

Establish a connection to the MIMIQ Services by loading the credentials from a JSON file.

Arguments

filename: file where to load the credentials (default: qperfect.json)

Note

The credentials are usually valid only for a small amount of time, so you may need to regenerate them from time to time.

Example

julia> savetoken("myqperfectcredentials.json")

julia> connection = loadtoken("myqperfectcredentials.json")

MimiqLink.refresh — Method

refresh(tokens, uri)

Refresh the tokens at the given uri / instance of MIMIQ.

MimiqLink.savetoken — Function

savetoken([filename][; uri="https://mimiq.qperfect.io/api"])

Establish a connection to the MIMIQ Services and save the credentials in a JSON file.

Arguments

filename: file where to save the credentials (default: qperfect.json)

Keyword arguments

uri: the uri of the MIMIQ Services (default: https://mimiq.qperfect.io/api)

Example

julia> savetoken("myqperfectcredentials.json")

julia> connection = loadtoken("myqperfectcredentials.json")