Execution utilities

MimiqCircuits.execute — Function

execute(connection, circuit[; kwargs...])
execute(connection, circuits[; kwargs...])

Execute a quantum circuit on the MIMIQ remote services.

The circuit is applied to the zero state and the resulting state is measured via sampling. Optionally amplitudes corresponding to few selected bit states (or bitstrings) can be returned from the computation.

Keyword Arguments

label::String: mnemonic name to give to the simulation, will be visible on the web interface

algorithm: algorithm to use by the compuation. By default "auto" will select the fastest algorithm between "statevector" or "mps".

nsamples::Integer: number of times to sample the circuit (default: 1000, maximum: 2^16)

bitstrings::Vector{BitString}: list of bit states to compute the amplitudes for (default: BitString[])

timelimit: number of minutes before the computation is stopped (default: maximum allowed or 30 minutes)

bonddim::Int: bond dimension for the MPS algorithm (default: 256, maximum: 4096)

entdim::Int: parameter to control pre compression of the circuit. Higher value makes simulations slower. (default: 16, minimum:4, maximum: 64)

seed::Int: a seed for running the simulation (default: random seed)

MimiqCircuits.getinput — Method

getinput(connection, execution)

Returns the first circuit and parameters for the given execution.

See also getinputs, getresults, getresult.

MimiqCircuits.getinputs — Method

getinputs(connection, execution)

Returns the circuits and parameters for the given execution.

See also getinput, getresults, getresult.

MimiqCircuits.getresult — Method

getresult(connection, execution)

Returns the first circuit and parameters for the given execution.

See also getinputs, getinput, getresults.

MimiqCircuits.getresults — Method

getresults(connection, execution[; interval=1])

Block until the given execution is finished and return the results.

Keyword Arguments

interval: time interval in seconds to check for job completion (default: 1)

See also getinputs, getinput, getresult.

MimiqCircuits.loadresults — Method

loadresults(file)

Load results from a given file.

MimiqCircuits.saveresults — Method

saveresults(file, results)

Save results to a given file.