Execution utilities

MimiqCircuits.execute — Method

execute(connection, circuit[; kwargs...])

Execute a quantum circuit on the MIMIQ remote services.

The circuit is applied to the zero state and the resulting state is measured via sampling. Optionally amplitudes corresponding to few selected bit states (or bitstrings) can be returned from the computation.

Keyword Arguments

label::String: mnemonic name to give to the simulation, will be visible on the web interface

algorithm: algorithm to use by the compuation. By default "auto" will select the fastest algorithm between "statevector" or "mps".

nsamples::Integer: number of times to sample the circuit (default: 1000, maximum: 2^16)

bitstrings::Vector{BitString}: list of bit states to compute the amplitudes for (default: BitString[])

timelimit: number of minutes before the computation is stopped (default: 5 minutes)

bonddim::Int: bond dimension for the MPS algorithm (default: 256, maximum: 4096)

entdim::Int: parameter to control pre compression of the circuit. Higher value makes simulations slower. (default: 16, minimum:4, maximum: 64)

seed::Int: a seed for running the simulation (default: random seed)

MimiqCircuits.executeqasm — Method

executeqasm(connection, qasmfilepath[, incfilepath[, incfilepath[, ...]]][; kwargs...])

Execute a quantum circuit on the MIMIQ remote services.

Optionally, additional include files can be provided to the computation. These will be used instead of the default ones provided (and cached) by the service. NOTE: Including standard files such as 'qelib1.inc' is not necessary, and might lead to slower computation.

The circuit is applied to the zero state and the resulting state is measured via sampling. Optionally amplitudes corresponding to few selected bit states (or bitstrings) can be returned from the computation.

Keyword Arguments

label::String: mnemonic name to give to the simulation, will be visible on the web interface

algorithm: algorithm to use by the compuation. By default "auto" will select the fastest algorithm between "statevector" or "mps".

nsamples::Integer: number of times to sample the circuit (default: 1000, maximum: 2^16)

bitstrings::Vector{BitString}: list of bit states to compute the amplitudes for (default: BitString[])

timelimit: number of minutes before the computation is stopped (default: 5 minutes)

bonddim::Int: bond dimension for the MPS algorithm (default: 256, maximum: 4096)

entdim::Int: parameter to control pre compression of the circuit. Higher value makes simulations slower. (default: 16, minimum:4, maximum: 64)

seed::Int: a seed for running the simulation (default: random seed)

MimiqCircuits.getinputs — Method

getinputs(connection, execution)

Returns the circuit and parameters for the given execution.

MimiqCircuits.getresults — Method

getresults(connection, execution; kwargs...)

Block until the given execution is finished and return the results.

Keyword Arguments

interval: time interval in seconds to check for job completion (default: 1)

MimiqCircuits.loadresults — Method

loadresults(file)

Load results from a given file.

MimiqCircuits.saveresults — Method

saveresults(file, results)

Save results to a given file.